refactor: 重构股票工具函数和配置管理模块
将股票相关工具函数合并到stockUtils模块
将配置管理拆分为vscodeConfig和staticConfig
新增lockManager模块处理封单监控逻辑
优化代码结构和提高可维护性
