+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - **封单监控**：新增股票涨跌停封单监控功能
   - 新增 `showLockCount` 配置项，开启后状态栏直接显示涨跌停封单金额
   - 新增 `enableLockTip` 配置项，开启后监控封单异动（涨停/跌停、开板、封单大幅变化等）并自动通知
+- 优化代码提高运行效率和可读性
 
 ## 1.5.0 2026-04-11
