+++ src/configs/staticConfig.js
@@ -1,0 +1,68 @@
+/**
+ * 项目静态配置
+ */
+
+/**
+ * 大盘指数代码
+ */
+const INDEX_CODES = [
+  "sh000001",
+  "sz399001",
+  "sz399006",
+  "sh000688",
+  "bj899050",
+];
+
+/**
+ * 行业板块代码
+ */
+const INDUSTRY_CODES = [
+  { code: "sh512880", name: "证券" },
+  { code: "sz159326", name: "电网设备" },
+  { code: "sh512400", name: "有色金属" },
+  { code: "sh512480", name: "半导体" },
+  { code: "sz159928", name: "消费" },
+  { code: "sz159206", name: "卫星" },
+  { code: "sh512690", name: "白酒" },
+  { code: "sh512010", name: "医药" },
+  { code: "sh515880", name: "通信" },
+  { code: "sh512660", name: "军工" },
+  { code: "sh515980", name: "人工智能" },
+  { code: "sh515220", name: "煤炭" },
+  { code: "sz159869", name: "游戏" },
+  { code: "sz159995", name: "芯片" },
+  { code: "sh516160", name: "新能源" },
+  { code: "sh562800", name: "稀有金属" },
+  { code: "sz159766", name: "旅游" },
+  { code: "sh512980", name: "传媒" },
+  { code: "sh515170", name: "食品饮料" },
+  { code: "sh512800", name: "银行" },
+  { code: "sh515230", name: "软件" },
+  { code: "sh561360", name: "石油" },
+  { code: "sh510230", name: "金融" },
+  { code: "sh512290", name: "生物医药" },
+  { code: "sz159516", name: "半导体设备" },
+  { code: "sh515210", name: "钢铁" },
+  { code: "sh512670", name: "国防" },
+  { code: "sh159625", name: "绿色电力" },
+  { code: "sh562500", name: "机器人" },
+  { code: "sz159611", name: "电力" },
+  { code: "sh560080", name: "中药" },
+  { code: "sz159870", name: "化工" },
+  { code: "sz159930", name: "能源" },
+  { code: "sz159992", name: "创新药" },
+  { code: "sh516970", name: "基建" },
+  { code: "sz159825", name: "农业" },
+  { code: "sh516150", name: "稀土" },
+  { code: "sz159865", name: "养殖" },
+  { code: "sz159755", name: "电池" },
+  { code: "sh516510", name: "云计算" },
+  { code: "sh515790", name: "光伏" },
+  { code: "sh512200", name: "房地产" },
+  { code: "sz159996", name: "家电" },
+  { code: "sz159227", name: "航空航天" },
+  { code: "sh516520", name: "智能驾驶" },
+  { code: "sh516620", name: "影视" },
+];
+
+module.exports = { INDEX_CODES, INDUSTRY_CODES };
